Move CommandBar into widgets/
