+++ include/fullscore/app_controller.h
@@ -7,9 +7,9 @@
 #include <fullscore/actions/action_base.h>
 #include <fullscore/actions/queue_action.h>
 #include <fullscore/models/reference_cursor.h>
+#include <fullscore/widgets/command_bar.h>
 #include <fullscore/widgets/gui_score_editor.h>
 #include <fullscore/widgets/measure_inspector.h>
-#include <fullscore/command_bar.h>
 #include <fullscore/follow_camera.h>
 #include <fullscore/playback_device_generic.h>
